Move validation helpers to dedicated packages
This is done to allow importing the logic in extensions without pulling
in all dependendies of the validation package. Actually, other projects
should never depend API validation packages. But we don't have a sync
out mechanism which filters out validation code, so here we are.
