+++ pkg/apis/core/validation/utils.go
@@ -110,27 +110,3 @@ func getPercentValue(intOrStringValue intstr.IntOrString) (int, bool) {
 	value, _ := strconv.Atoi(intOrStringValue.StrVal[:len(intOrStringValue.StrVal)-1])
 	return value, true
 }
-
-// ShouldEnforceImmutability compares the given slices and returns if a immutability should be enforced.
-// It mainly checks if the order of the same elements in `new` and `old` is the same, i.e. only an addition
-// of elements to `new` is allowed.
-func ShouldEnforceImmutability(new, old []string) bool {
-	sizeDelta := len(new) - len(old)
-	if sizeDelta > 0 {
-		newA := new[:len(new)-sizeDelta]
-		if equal(newA, old) {
-			return false
-		}
-		return ShouldEnforceImmutability(newA, old)
-	}
-	return sizeDelta < 0 || sizeDelta == 0
-}
-
-func equal(new, old []string) bool {
-	for i := range new {
-		if new[i] != old[i] {
-			return false
-		}
-	}
-	return true
-}
